Refactor DatasetConfig into its own file.
